requirements.txt
git+https://github.com/openai/whisper.git
openai
python-dotenv